+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 2. Created Folder Structure
 3. Initialised npm & installed inquirer & jest packages
 4. Created questions.js in the lib folder and created the questions array
+5. Created buildLogo.js in the lib folder to hold code to control logo build
+6. Add code to ask questions and write the Logo to file